+++ examples/resources/clearpass_extension_instance_config/import.sh
@@ -1,0 +1,2 @@
+# Extension instance config can be imported using the extension instance ID
+terraform import clearpass_extension_instance_config.intune_config 0e57dcbd-aefe-47c3-8e25-ec8c320477d5